Final refactoring changs
List of changes:
1. Change the `YEAR` cli argument to only accept integers
2. Changed the output of `capgains calc` when there are no transactions
and no capital gains for a given ticker
3. Changed the exchange rate method where it fetches the closest
preceeding date to use a more efficient algorithm rather than just
subtracting the date by one and checking if the rate for that day
exists
4. Fixed bug where we were subtracting the commission for a BUY
transaction. The commission should be added since it is an added cost
for buying the share.
5. Removed `to_list()` method from Transaction class. Instead I created
a test helper since this method was only used for aiding in testing.
6. Made a common exchange rate mocker to share across tests, instead of
having a unique one for every test. This one just takes the transactions
fixture and generates fake exchange rate values for the transactions.
7. Changed the values in the transactions so that the numbers are easier
to work with when manually verifying the capital gains + acb.