Adding a currency column to the expected input transactions (#7)
Fixed #5

Now, we expect a currency to be added next to every transaction that specifies the currency used in the 'Price' and 'Commission' column.

The calculated values of the calc command will still be displayed in CAD though.

Breakdown of the changes:

1. capgains_calc.py:
a. First map currency to a tuple of (min_date, max_date) where min_date is the earliest date a transaction using this currency occured, and max_date is the latest.

So you might end up with something that looks like:
```
{
    "CAD" : (January 1st 2020, January 2nd 2020),
    "USD" : (January 1st 2020, February 1st 2020)
}
```

b. Go over all the currencies detected in step a, create an ExchangeRate object using the min_date and max_date from the tuples created in step b, and map each currency to this new ExchangeRate object.

This gives you something like looks like:
```
{
    "CAD" : (ExchangeRate object with rates from Jan-1-2020 -> Jan-2-2020),
    "USD" : (ExchangeRate object with rates from Jan-1-2020 -> Feb-1-2020) 
}
```

2. exchange_rate.py:
a. The constructor will check if the currency_from variable is 'CAD', and initialize self._rates to just set every exchange rate to 1 for all the dates. This is because the API cannot do a get request for CAD->CAD, and the rate is just 1 since its the same currency. If the currency is not CAD, then it will proceed to do the original initialization which gets the values from the API call to Bank of Canada.