Increase test coverage to 100% (#11)
Test coverage is now 100%. Also, the codecov.yml file was created to enforce 100% line coverage for all future PRs.

Only changes to code that need to be made were in capgains_calc.py. There was 1 single line that wasn't being tested, but it was an impossible case since we expected transactions to be in order anyways. So, I rewrote the function and simplified it so that there isn't that giant for-loop anymore. Now, we just generate a group of lists with common currencies, and we know the order of the transactions are correct, so we can assume the first is the min date and the last is the max date.