Use a cleaner table format for output
The old table format used for the command outputs was clean but it was
hard to distinguish between the table and information outside the table
(like total capital gains). I changed the table format to use `psql`
format which adds borders around the table headers.

Also included a minor change where I changed the poetry package name to
cad-capgains. This is the name that people will use when downloading the
package off PyPi. The CLI command still remains as 'capgains', as does
the python module.