Add support for BoC noon rates
Add support for fetching Bank of Canada noon rates. This allows us to
set the minimum date for transactions to be May 1, 2007 instead of Jan.
3, 2017.

For transactions with dates that fall before Jan. 3, 2017, they will use
the noon rate from Bank of Canada. For transactions with dates that fall
after or on Jan. 3, 2017, they will use the indicative rate from Bank of
Canada.