Add support for BoC noon rates (#15)
* Add support for BoC noon rates

Add support for fetching Bank of Canada noon rates. This allows us to
set the minimum date for transactions to be May 1, 2007 instead of Jan.
3, 2017.

For transactions with dates that fall before Jan. 3, 2017, they will use
the noon rate from Bank of Canada. For transactions with dates that fall
after or on Jan. 3, 2017, they will use the indicative rate from Bank of
Canada.

* Update README

Update the README.md so that it reflects the fact that we now support
transactions from 2007 and onwards.

Co-authored-by: Eddy Maric <eddy.maric98@gmail.com>