Support partial share purchases/sales (#25)
* Support partial share purchases/sales

Something that is possible through some brokerages is the purchase and
sale of partial shares. This means that the `qty` value isn't
necessarily an integer, but could be represented as a floating point.
This can also occur for things like bitcoin or currency, where you will
most likely hold a certain fraction of a coin/currency. The primary
change was to support having decimal/float values for the quantity.

Additionally, I noticed that we were doing a lot of floating point math.
While this is good for approximations, it is better to be precise. I
discovered that there is a `decimal` python library that allows you to
work with precise decimal objects that don't have weird rounding rules
like floats. Anywhere where we did floating point math or where we
returned floats, I converted to use a Decimal instead.

* Change test names + Update README

Change test names so that they are named appropriately. We are now
checking that qty/commissions/price are valid numbers, not simply just
integers/floats.

Update README to include the fact that we support fractional shares.